[C8] Split extrude.go by responsibility (#2235)
model/feature/extrude.go was 524 LOC, over the 500-line ceiling. Split the
extrude feature along its responsibilities; each file is now under the ceiling:

- extrude.go (205): the ExtrudeFeature (definition, accessors, Recompute, the
  sheet/surface variant recomputeSurface, buildTool, resolveProfiles) and the
  ExtrudeFeatures adder collection.
- extrude_combine.go (119): the merge/dissolve policy — combine (the boolean
  op), combinePrisms (per-region), SetExtrudeDissolve (abutting-prism dissolve
  toggle) and appendCombined (running-body accumulation).
- extrude_prism.go (232): the prism builder — the capped/uncapped extrusion
  shell, tapered/offset loops, profile sheets, and cap/side face construction.

Behaviour-preserving: split programmatically by receiver-qualified
declaration, verified every non-comment code line is byte-identical (only
import redistribution + gofmt one-liner alignment differ, the latter
restored); imports resolved per file with goimports. Feature package builds,
archguard passes, lint 0. Net delta of recognizers / tolerance constants /
fallback sites / type assertions: 0.

Note: the prism builder's cap/side edge construction is also flagged C6-a
(guarded edge reuse); that guarding is the separate phase-2 task.

Part of #2183 (C8: structure/size limits). Closes #2235.

(cherry picked from commit 5d11a7481c1afa0f332a183b45c64d1d80198b38)
